1. Secure Notes - Android Application
The source code contained within this file group is that used to create a basic Android Application using the programming languages Kotlin and Java.

This project was an early creation, created as a final piece of work for a short Computer Science course undertaken before enrolling on my MSc Computer Science with Artificial Intelligence. This file involved a great deal of self learning as the particular course undertaken, did not in fact teach Kotlin or Java languages and instead focused on Web Application development using Python, HTML and CSS.