2. Finance Web Application
This file contains the source code for a Web Application that was created as part of a Computer Science course.

Within the file itself a number of Source Code files can be found. With the exception of Helpers.py and apology.py which were provided by the course facilitators, the remaining files were written and completed myself.

The Web Application works by using an API Key to access and query the data of IEX in order to obtain financial market data via get requests. It then uses this data in order to simulate a Web Application whereby stocks can be purchased and sold in order to maintain a Portfolio which is managed within a SQL Database using SQLite3.

Each HTML file has its own .py file in order to control its logic.