2. Finance Web Application
This file contains the Source Code for a Web Application written in Python, HTML and CSS, completed as part of a Computer Science course.

With the exception of helpers.py and apology.py, which were provided by the course facilitators, the Source Code files were created by myself.

The Web Application works by using an API Key in order to draw financial market data from IEX so that this can be used in order to simulate a Web Application for the building of a Stock Portfolio. This allows the users to purchase, sell  and build a fictional portfolio of Stocks which are stored using a SQL Databased using SQLite3.